modules/location_producer/requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
kafka-python==2.0.2